chore: improve Textarea exports
