+++ packages/chakra-ui-core/src/Textarea/index.js
@@ -1,34 +1,2 @@
-import styleProps from '../config/props'
-import { inputProps } from '../Input/input.props'
-import Input from '../Input'
-import { forwardProps } from '../utils'
-
-const Textarea = {
-  name: 'Textarea',
-  model: {
-    prop: 'inputValue',
-    event: 'change'
-  },
-  props: {
-    ...styleProps,
-    ...inputProps,
-    inputValue: String
-  },
-  render (h) {
-    return h(Input, {
-      props: {
-        ...forwardProps(this.$props),
-        as: 'textarea',
-        py: '8px',
-        minHeight: '80px',
-        fontFamily: 'body',
-        lineHeight: 'shorter'
-      },
-      on: {
-        input: (value, $e) => this.$emit('change', value, $e)
-      }
-    }, this.$slots.default)
-  }
-}
-
+import Textarea from './Textarea'
 export default Textarea
